Move bufsize_to_datatype to utils
