+++ src/hdf5_init.c
@@ -1,4 +1,5 @@
 #include "hdf5_init.h"
+#include "utils.h"
 #include <hdf5.h>
 #include <stdio.h>
 #include <stdlib.h>
@@ -40,34 +41,6 @@ void initialize_file(char *filename, hid_t *file_id, hid_t fapl, hid_t fcpl)
     fprintf(stderr, "Error in creating file_id in intialize_file\n");
     return;
   }
-}
-
-hid_t bufsize_to_datatype(int dtype)
-{
-  hid_t datatype;
-  switch (dtype) {
-    case 1: {
-      datatype = H5T_STD_U8LE;
-      break;
-    }
-    case 2: {
-      datatype = H5T_STD_U16LE;
-      break;
-    }
-    case 4: {
-      datatype = H5T_STD_U32LE;
-      break;
-    }
-    case 8: {
-      datatype = H5T_STD_U64LE;
-      break;
-    }
-    default: {
-      fprintf(stderr, "Error in datatype, please check input dtype\n");
-      return;
-    }
-  }
-  return datatype;
 }
 
 void create_merlin_dataset(hid_t *merlin_dataset_id,
