+++ src/utils.c
@@ -121,3 +121,31 @@ restore:
       current_file, __LINE__);
   }
 }
+
+hid_t bufsize_to_datatype(int dtype)
+{
+  hid_t datatype;
+  switch (dtype) {
+    case 1: {
+      datatype = H5T_STD_U8LE;
+      break;
+    }
+    case 2: {
+      datatype = H5T_STD_U16LE;
+      break;
+    }
+    case 4: {
+      datatype = H5T_STD_U32LE;
+      break;
+    }
+    case 8: {
+      datatype = H5T_STD_U64LE;
+      break;
+    }
+    default: {
+      fprintf(stderr, "Error in datatype, please check input dtype\n");
+      return;
+    }
+  }
+  return datatype;
+}
